moved numpy compatibility stuff to a new file
