+++ src/prpy/__init__.py
@@ -32,3 +32,4 @@ import base, dependency_manager, logger, ik_ranking, planning, simulation, tsr, 
 from named_config import ConfigurationLibrary
 from clone import Clone, Cloned
 from bind import bind_subclass
+import compatibility
