+++ src/prpy/ik_ranking.py
@@ -28,20 +28,7 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-from distutils.version import LooseVersion
 import numpy
-
-# handle older versions of numpy without norm axis argument
-if LooseVersion(numpy.__version__) < LooseVersion('1.8.0'):
-    oldnorm = numpy.linalg.norm
-    def newnorm(x, ord=None, axis=None):
-        if axis is None:
-            return oldnorm(x, ord=ord)
-        elif isinstance(axis, (int,long)):
-            return numpy.apply_along_axis(oldnorm, axis, x)
-        else:
-            raise NotImplementedError('older numpy without norm axis')
-    numpy.linalg.norm = newnorm
 
 def NoRanking(robot, ik_solutions):
     """
